move HomePanel to homePanel directory
